+++ apps/blade/src/app/admin/_components/SchoolBarChart.tsx
@@ -10,7 +10,7 @@ import {
 } from "recharts";
 
 import type { ChartConfig } from "@forge/ui/chart";
-import { SCHOOLS } from "@forge/consts/knight-hacks";
+import type { SCHOOLS } from "@forge/consts/knight-hacks";
 import { Card, CardContent, CardHeader, CardTitle } from "@forge/ui/card";
 import { ChartContainer, ChartTooltip } from "@forge/ui/chart";
