+++ apps/blade/src/app/dashboard/_components/hacker-dashboard/hacker-dashboard.tsx
@@ -18,7 +18,7 @@ export default async function HackerDashboard({
 }) {
   const [resume, pastHackathons] = await Promise.allSettled([
     api.resume.getResume(),
-    api.hacker.getHackathons(),
+    api.hackathon.getPastHackathons(),
   ]);
 
   return (
